+++ Elements/src/Spatial/CellComplex/Cell.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Elements;
+using Elements.Geometry.Solids;
 using Elements.Geometry;
 using Newtonsoft.Json;
 
@@ -10,7 +10,7 @@ namespace Elements.Spatial.CellComplex
     /// <summary>
     /// A cell: a 3-dimensional closed cell within a complex
     /// </summary>
-    public class Cell : CellChild
+    public class Cell : CellChild<Extrude>
     {
         /// <summary>
         /// Bottom face. Can be null. Expected to be duplicated in list of faces.
@@ -26,6 +26,9 @@ namespace Elements.Spatial.CellComplex
         /// All faces
         /// </summary>
         public List<long> FaceIds;
+
+        [JsonIgnore]
+        private Extrude _geometry;
 
         /// <summary>
         /// Represents a unique Cell wtihin a CellComplex.
@@ -46,6 +49,12 @@ namespace Elements.Spatial.CellComplex
             {
                 this.TopFaceId = topFace.Id;
             }
+            if (bottomFace != null && topFace != null)
+            {
+                var bottom = this.GetBottomFace().GetGeometry();
+                var top = this.GetTopFace().GetGeometry();
+                this._geometry = new Extrude(bottom, top.Centroid().Z - bottom.Centroid().Z, new Vector3(0, 0, 1), false);
+            }
             this.FaceIds = faces.Select(ds => ds.Id).ToList();
         }
 
@@ -58,6 +67,15 @@ namespace Elements.Spatial.CellComplex
             this.FaceIds = faceIds;
             this.BottomFaceId = bottomFaceId;
             this.TopFaceId = topFaceId;
+        }
+
+        /// <summary>
+        /// Get a Solid representing this Geometry
+        /// </summary>
+        /// <returns></returns>
+        public override Extrude GetGeometry()
+        {
+            return this._geometry;
         }
 
         /// <summary>
@@ -88,12 +106,98 @@ namespace Elements.Spatial.CellComplex
         }
 
         /// <summary>
+        /// Get the closest associated face to the supplied position
+        /// </summary>
+        /// <param name="position"></param>
+        /// <returns></returns>
+        public Face GetClosestFace(Vector3 position)
+        {
+            return this.GetFaces().OrderBy(f => position.DistanceTo(f.GetGeometry())).ToList().First();
+        }
+
+        /// <summary>
         /// Get associated Segments
         /// </summary>
         /// <returns></returns>
         public List<Segment> GetSegments()
         {
             return this.GetFaces().Select(face => face.GetSegments()).SelectMany(x => x).Distinct().ToList();
+        }
+
+        /// <summary>
+        /// Get the closest associated segment to the supplied position
+        /// </summary>
+        /// <param name="position"></param>
+        /// <returns></returns>
+        public Segment GetClosestSegment(Vector3 position)
+        {
+            return this.GetSegments().OrderBy(s => position.DistanceTo(s.GetGeometry())).ToList().First();
+        }
+
+
+        /// <summary>
+        /// Get associated Vertices
+        /// </summary>
+        /// <returns></returns>
+        public List<Vertex> GetVertices(Vector3? point = null)
+        {
+            return this.GetFaces().Select(f => f.GetVertices()).SelectMany(x => x).Distinct().ToList();
+        }
+
+        /// <summary>
+        /// Get the closest associated vertex to the supplied position
+        /// </summary>
+        /// <param name="position"></param>
+        /// <returns></returns>
+        public Vertex GetClosestVertex(Vector3 position)
+        {
+            return this.GetVertices().OrderBy(v => position.DistanceTo(v.Value)).ToList().First();
+        }
+
+        /// <summary>
+        /// Get list of Cells that are neighbors
+        /// </summary>
+        /// <returns></returns>
+        public List<Cell> GetNeighbors()
+        {
+            return this.GetFaces().Select(f => f.GetCells()).SelectMany(x => x).Distinct().Where(c => c.Id != this.Id).ToList();
+        }
+
+        /// <summary>
+        /// Get a list of the neighbor that shares a specific Face with this cell.
+        /// Can be null if the face is not shared.
+        /// </summary>
+        /// <param name="face">Shared face</param>
+        /// <returns></returns>
+        public Cell GetNeighbor(Face face)
+        {
+            if (face == null)
+            {
+                return null;
+            }
+
+            var cells = face.GetCells().Where(cell => cell.Id != this.Id).ToList();
+
+            if (cells.Count == 0)
+            {
+                return null;
+            }
+            if (cells.Count > 1)
+            {
+                throw new Exception("Cells should only have one neighbor which share this face");
+            }
+
+            return cells[0];
+        }
+
+        /// <summary>
+        /// Get the closest associated cell to the supplied position, calculating to the approximate center of the cell
+        /// </summary>
+        /// <param name="position"></param>
+        /// <returns></returns>
+        public Cell GetClosestCell(Vector3 position)
+        {
+            return this.GetNeighbors().OrderBy(cell => cell.DistanceTo(position)).ToList().First();
         }
     }
 }